Fixes transportation issues on MacOS
The issue with transports on MacOS is that the client movement info is
not updated on MSG_MOVE_WORLDPORT_ACK, instead, the information must be
sent then to avoid being teleported elsewhere. Sending a heartbeat will
inform the client of the new position which is set after a teleport
event is scheduled.

Another issue this leads to is the client player camera facing south on
a successful port. The movement of the transport object does not affect
the camera at all. The client resets the camera in two situations as
I've discovered them: Upon leaving a transport object -and- triggering a
SMSG_STANDSTATE_UPDATE event. This fix sends the event with the current
state such that only the camera is reset.